This exercises an edge case that fails in real world.
When the user wraps an interation in an arrow function, when transformed
it fails. I'm not sure if it's because this transformation tries to put
multiple lines in an arrow function without braces, which is invalid or what.

I think that arrow functions wrapping interation should be removed as now they
are not needed